split out job_result
