+++ lib/salesforce-bulk-query.rb
@@ -6,4 +6,5 @@ require 'salesforce-bulk-query/version'
 require 'salesforce-bulk-query/client'
 require 'salesforce-bulk-query/connection'
 require 'salesforce-bulk-query/job'
+require 'salesforce-bulk-query/job_result'
 require 'salesforce-bulk-query/errors'
